// dashboard/src/legacy/main/home/app-dashboard/app-view/tabs/activity-feed/events/types.ts
export const porterAppEventValidator = z.discriminatedUnion("type", [
  z.object({
    id: z.string(),
    created_at: z.string(),
    updated_at: z.string(),
    status: z.string().optional().default(""),
    type: z.literal("BUILD"),
    type_external_source: z.string().optional().default(""),
    porter_app_id: z.number(),
    metadata: porterAppBuildEventMetadataValidator,
  }),
  z.object({
    id: z.string(),
    created_at: z.string(),
    updated_at: z.string(),
    status: z.string().optional().default(""),
    type: z.literal("DEPLOY"),
    type_external_source: z.string().optional().default(""),
    porter_app_id: z.number(),
    metadata: porterAppDeployEventMetadataValidator,
  }),
  z.object({
    id: z.string(),
    created_at: z.string(),
    updated_at: z.string(),
    status: z.string().optional().default(""),
    type: z.literal("PRE_DEPLOY"),
    type_external_source: z.string().optional().default(""),
    porter_app_id: z.number(),
    metadata: porterAppPreDeployEventMetadataValidator,
  }),
  z.object({
    id: z.string(),
    created_at: z.string(),
    updated_at: z.string(),
    status: z.string().optional().default(""),
    type: z.literal("INITIAL_DEPLOY"),
    type_external_source: z.string().optional().default(""),
    porter_app_id: z.number(),
    metadata: porterAppPreDeployEventMetadataValidator,
  }),
  z.object({
    id: z.string(),
    created_at: z.string(),
    updated_at: z.string(),
    status: z.string().optional().default(""),
    type: z.literal("APP_EVENT"),
    type_external_source: z.string().optional().default(""),
    porter_app_id: z.number(),
    metadata: porterAppAppEventMetadataValidator,
  }),
  z.object({
    id: z.string(),
    created_at: z.string(),
    updated_at: z.string(),
    type: z.literal("NOTIFICATION"),
    porter_app_id: z.number(),
    metadata: porterAppNotificationEventMetadataValidator,
  }),
  z.object({
    id: z.string(),
    created_at: z.string(),
    updated_at: z.string(),
    type: z.literal("AUTO_ROLLBACK"),
    type_external_source: z.string().optional().default(""),
    porter_app_id: z.number(),
    metadata: porterAppDeployEventMetadataValidator,
  }),
]);

export const getPorterAppEventsValidator = z
  .array(porterAppEventValidator)
  .optional()
  .default([]);

export type PorterAppEvent = z.infer<typeof porterAppEventValidator>;
export type PorterAppBuildEvent = PorterAppEvent & { type: "BUILD" };
export type PorterAppDeployEvent = PorterAppEvent & { type: "DEPLOY" };
export type PorterAppPreDeployEvent = PorterAppEvent & { type: "PRE_DEPLOY" };
export type PorterAppInitialDeployEvent = PorterAppEvent & {
  type: "INITIAL_DEPLOY";
};